torpedoes code update
