+++ torpedoes.py
@@ -1,6 +1,9 @@
 import pigpio
 import time
 from logpy.LogPy import Logger
+from ports import TORPEDO_FIRE_DRIVER_PORT, TORPEDO_READY_CLIENT_PORT
+from rov_comm import Client
+from torpedo_handling import TorpedoHandling
 
 PWM_serwo = 23
 reed_switch = 24
@@ -13,6 +16,32 @@ TIME_OF_INITIAL_ROTATION = 0.5 # time required to make one full rotation with ma
 FAST_ROTATION = 230 # you add/sub this to/from the neutral position
 SLOW_ROTATION = 70 # same as above
 NEUTRAL = 458
+
+class Torpedoes:
+    is_ready = True
+    def __init__(self):
+        self.client_fire = Client(TORPEDO_FIRE_DRIVER_PORT)
+        self.client_ready = Client(TORPEDO_READY_CLIENT_PORT)
+        self.torpedo_handling = TorpedoHandling()
+
+    def fire(self):
+        command = self.client_fire.get_data()
+        #print(command)
+        if command == "FIRE" and self.is_ready:
+            self.torpedo_handling.sequence()
+
+            self.client_ready.send_data('WAIT')
+            self.is_ready = False
+            self.reload_torpedo()
+
+    def reload_torpedo(self):
+        reload_time = 2   # czas przeładowywania, do ustawienia
+        """
+        TUTAJ PRZEŁADOWANIE
+        """
+        time.sleep(reload_time)  # jak umiecie jakieś lepsze czekanie, to zmieńcie
+        self.client_ready.send_data('READY')
+        self.is_ready = True
 
 
 class TorpedoHandling:
